split AllItems to ItemBuilder ModelBuilder
