+++ src/main/java/shagejack/industrimania/content/contraptions/ore/ItemOreChunk.java
@@ -1,12 +1,6 @@
 package shagejack.industrimania.content.contraptions.ore;
 
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import shagejack.industrimania.content.worldGen.record.OreType;
-import shagejack.industrimania.registers.AllItems;
-
-import java.util.List;
-import java.util.Map;
 
 public class ItemOreChunk extends Item {
