+++ src/main/java/shagejack/industrimania/content/metallurgyAge/block/smeltery/ironOreSlag/IronOreSlagTileEntity.java
@@ -10,7 +10,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import shagejack.industrimania.content.metallurgyAge.item.smeltery.cluster.IronCluster;
 import shagejack.industrimania.foundation.tileEntity.SmartTileEntity;
 import shagejack.industrimania.foundation.tileEntity.TileEntityBehaviour;
-import shagejack.industrimania.registers.AllItems;
+import shagejack.industrimania.registers.item.AllItems;
 import shagejack.industrimania.registers.AllTileEntities;
 
 import java.util.List;
